Better ref fowarding
