--- react/components/EXPERIMENTAL_Table/DataTable/RadioactiveTableParts.tsx
@@ -1,4 +1,0 @@
-import Row from './Row'
-
-export { DataTable, Tbody, Thead } from './index'
-export { Row }
